# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  models.ScanStatus:
    enum:
    - pending
    - running
    - stopped
    - failed
    - completed
    type: string
    x-enum-varnames:
    - StatusPending
    - StatusRunning
    - StatusStopped
    - StatusFailed
    - StatusCompleted
  models.ScanTaskResponse:
    properties:
      batch_id:
        type: string
      error:
        type: string
      progress:
        type: integer
      result:
        items:
          additionalProperties: true
          type: object
        type: array
      started_at:
        description: RFC3339 format
        type: string
      status:
        $ref: '#/definitions/models.ScanStatus'
      target:
        type: string
      task_id:
        type: string
      updated_at:
        description: RFC3339 format
        type: string
      user_id:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact:
    name: API Support
  description: Security Scanning API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Napscan API
  version: "1.0"
paths:
...
  /ffuf/scan/{task_id}/report:
    get:
      description: Returns the scan report/results. Only available when status is
        'completed' (works for all scanners)
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              result:
                items:
                  additionalProperties: true
                  type: object
                type: array
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get Scan Report
      tags:
      - Scan Control
  /ffuf/scan/{task_id}/status:
    get:
      description: Returns the unified status and progress of a scan task (works for
        all scanners)
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ScanTaskResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get Scan Status
      tags:
      - Scan Control
  /ffuf/scan/{task_id}/stop:
    post:
      description: Stops a running or pending scan task by task ID (works for all
        scanners)
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ScanTaskResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Stop Scan
      tags:
      - Scan Control
...
  /nmap/scan/{task_id}/report:
    get:
      description: Returns the scan report/results. Only available when status is
        'completed' (works for all scanners)
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              result:
                items:
                  additionalProperties: true
                  type: object
                type: array
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get Scan Report
      tags:
      - Scan Control
  /nmap/scan/{task_id}/status:
    get:
      description: Returns the unified status and progress of a scan task (works for
        all scanners)
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ScanTaskResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get Scan Status
      tags:
      - Scan Control
  /nmap/scan/{task_id}/stop:
    post:
      description: Stops a running or pending scan task by task ID (works for all
        scanners)
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ScanTaskResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Stop Scan
      tags:
      - Scan Control
...
  /sslyze/scan/{task_id}/report:
    get:
      description: Returns the scan report/results. Only available when status is
        'completed' (works for all scanners)
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              result:
                items:
                  additionalProperties: true
                  type: object
                type: array
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get Scan Report
      tags:
      - Scan Control
  /sslyze/scan/{task_id}/status:
    get:
      description: Returns the unified status and progress of a scan task (works for
        all scanners)
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ScanTaskResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get Scan Status
      tags:
      - Scan Control
  /sslyze/scan/{task_id}/stop:
    post:
      description: Stops a running or pending scan task by task ID (works for all
        scanners)
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ScanTaskResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Stop Scan
      tags:
      - Scan Control
...
  /zap/scan:
    post:
      consumes:
      - application/json
      description: Run ZAP Spider and Active Scan
      parameters:
      - description: Target URL
        in: body
        name: target
        required: true
        schema:
          properties:
            batch_id:
              type: string
            target:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Start ZAP Scan
      tags:
      - ZAP
  /zap/scan/{task_id}/report:
    get:
      description: Returns the scan report/results. Only available when status is
        'completed' (works for all scanners)
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              result:
                items:
                  additionalProperties: true
                  type: object
                type: array
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get Scan Report
      tags:
      - Scan Control
  /zap/scan/{task_id}/status:
    get:
      description: Returns the unified status and progress of a scan task (works for
        all scanners)
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ScanTaskResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get Scan Status
      tags:
      - Scan Control
  /zap/scan/{task_id}/stop:
    post:
      description: Stops a running or pending scan task by task ID (works for all
        scanners)
      parameters:
      - description: Task ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ScanTaskResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Stop Scan
      tags:
      - Scan Control
  /zap/scan/async:
    post:
      consumes:
      - application/json
      description: Start a ZAP scan asynchronously. Returns task_id immediately.
      parameters:
      - description: Scan Request
        in: body
        name: request
        required: true
        schema:
          properties:
            batch_id:
              type: string
            target:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ScanTaskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Start ZAP Scan (Async)
      tags:
      - ZAP
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"